weixin_outer/requirement.txt
asgi-redis==1.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.7.1
Automat==0.6.0
certifi==2017.4.17
channels==1.1.6
chardet==3.0.4
constantly==15.1.0
cssselect==1.0.1
daphne==1.3.0
Django==1.10
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
hyperlink==17.3.0
ibm-db==2.0.7
ibm-db-django==1.1.0.2
ibm-db-sa==0.3.3
idna==2.5
incremental==17.5.0
lxml==3.8.0
msgpack-python==0.4.8
pycrypto==2.6.1
pyquery==1.2.17
redis==2.10.5
requests==2.18.1
six==1.10.0
SQLAlchemy==1.1.12
Twisted==17.5.0
txaio==2.8.1
urllib3==1.21.1
zope.interface==3.8.0
